Models/Brain_tumor/requirements.txt
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
pillow==10.0.0
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2